+++ src/main/java/entity/Product.java
@@ -1,14 +1,14 @@
 package entity;
 
 public class Product {
-    private int id;
+    private long id;
     private String productName;
     private float price;
     private float weight;
     private String color;
     private int productCount;
 
-    public Product(int id, String productName, float price, float weight, String color, int productCount) {
+    public Product(long id, String productName, float price, float weight, String color, int productCount) {
         this.id = id;
         this.productName = productName;
         this.price = price;
@@ -17,7 +17,7 @@ public class Product {
         this.productCount = productCount;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
